+++ lionagi/operations/instruct/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2023-2026, HaiyangLi <quantocean.li at gmail dot com>
-# SPDX-License-Identifier: Apache-2.0
-
-from .instruct import instruct
-
-__all__ = ["instruct"]